Change below: + split server on server/microServers + command run server now with npm start + change hit API server on client/components/board + change hit API server on client/pages/Game
