--- client/src/config/socket.js
@@ -1,3 +1,0 @@
-import io from "socket.io-client";
-
-export const socket = io("http://localhost:9000/");
