+++ server/package.json
@@ -5,13 +5,15 @@
   "main": "index.js",
   "scripts": {
     "dev": "NODE_ENV=development nodemon app.js",
-    "test": "jest --runInBand --detectOpenHandles --forceExit"
+    "test": "jest --runInBand --detectOpenHandles --forceExit",
+    "start": "concurrently 'nodemon app.js' 'nodemon microServers/boardSocket.js' 'nodemon microServers/webRtcSocket.js'"
   },
   "keywords": [],
   "author": "Bima, Fatur, Wiby, Hika",
   "license": "ISC",
   "dependencies": {
     "bcrypt": "^5.0.0",
+    "concurrently": "^5.3.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
